+++ app/src/main/java/uk/ac/soton/adDashboard/records/Impression.java
@@ -3,22 +3,20 @@ package uk.ac.soton.adDashboard.records;
 
 import java.time.LocalDateTime;
 
-import uk.ac.soton.adDashboard.enums.Context;
 import uk.ac.soton.adDashboard.enums.Gender;
-import uk.ac.soton.adDashboard.enums.Income;
 
 public class Impression extends LogRow {
 
   private final long id; //also the key of the hashmap :)
 
-  private final int cost;
+  private final double cost;
 
 
   private final LocalDateTime date;
 
   private Gender gender;
 
-  public Impression(String date, long id, int cost)
+  public Impression(String date, long id, double cost)
       throws Exception {
     super(id);
     this.date = parseDateTime(date);
@@ -33,7 +31,7 @@ public class Impression extends LogRow {
   }
 
 
-  public int getCost() {
+  public double getCost() {
     return cost;
   }
